Nectar 0.2: original JAudio engine and 120 FPS option
Audio now runs the game's own JAudio engine instead of a reimplementation.
The sequencer, instrument banks, oscillators and channel manager come from
the decompilation unchanged; only the GameCube DSP boundary is replaced by a
software renderer, adapted from pikmin-nextos and driven through SDL2.
Music, sound effects and cinematic audio all play.

Enabled with -DPIKMIN_NATIVE_JAUDIO=ON, which the Linux packaging script now
passes. With the option off the previous native mixer is used and nothing
changes. Verified by --audio-self-test across scenes, stages, boss
transitions, muting, sound effects and cinematic streams.

Frame rate: a 120 FPS option joins 30 and 60. The game integrates by elapsed
time, so a higher tick rate does not speed up gameplay.

Launcher: the installer looked for pikmin/pikmin.real while the build has
produced nectar/nectar.real since the rename, so installation failed. The
package wrappers also exported NECTAR_EXECUTABLE_PATH while the launcher read
PIKMIN_EXECUTABLE_PATH, leaving it unable to locate its own package. Disc
names from the FST are now decoded explicitly rather than through the active
locale, which aborted extraction on the disc's Shift-JIS filenames.

Docs: AI_HANDOFF.md claimed 30 FPS gameplay and contradicted SIGUIENTE_SESION.
It now records how the frame rate actually works, and that the TIME-002
interpolation machinery is built but never invoked.

+++ src/jaudio/hvqm_play.c
@@ -1,4 +1,5 @@
 #include "jaudio/hvqm_play.h"
+#include <stdint.h>
 #include "Dolphin/os.h"
 #include "hvqm4.h"
 #include "jaudio/dspbuf.h"
@@ -108,7 +109,7 @@ static void __ReLoad()
 		dvd_active += 1;
 
 		int num_bufs = 3;
-		DVDT_LoadtoDRAM(dvdcount, filename, (u32)dvd_buf[dvdcount % num_bufs], dvdcount << 0x13, dvdload_size, NULL, __LoadFin);
+		DVDT_LoadtoDRAM(dvdcount, filename, (uintptr_t)dvd_buf[dvdcount % num_bufs], dvdcount << 0x13, dvdload_size, NULL, __LoadFin);
 		OSRestoreInterrupts(inter);
 	}
 }
@@ -234,7 +235,7 @@ void Jac_HVQM_Init(immut char* movieFilePath, u8* data, u32 bufferSize)
 	dvdfile_size   = DVDT_CheckFile(movieFilePath);
 	dvdfile_size -= 0x80000;
 	volatile u32 status;
-	DVDT_LoadtoDRAM(dvdcount, movieFilePath, (u32)dvd_buf[dvdcount % 3], 0, 0x80000, (u32*)&status, 0);
+	DVDT_LoadtoDRAM(dvdcount, movieFilePath, (uintptr_t)dvd_buf[dvdcount % 3], 0, 0x80000, (u32*)&status, 0);
 	while (status == 0) { }
 
 	dvd_ctrl[0].mFileOffset = 0;
